% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcIntraInter}
\alias{calcIntraInter}
\title{Calculate Intra- and Interclass Correlation Coefficients}
\usage{
calcIntraInter(
  fmddf,
  exclOneMT = FALSE,
  ther.sub = NULL,
  lmer = TRUE,
  verbose = TRUE,
  fnOut = "results_stat",
  toXls = TRUE
)
}
\arguments{
\item{fmddf}{A dataframe containing FMD data.}

\item{exclOneMT}{Logical, if additional groupings with one muscle test
excluded should be performed. Defaults to FALSE.}

\item{ther.sub}{NULL or List. If left at the default NULL, no additional
splitting with therapist subgroups is performed. Provide a list with
n therapist IDs in each list element to calculate additional inter-rater
correlation coefficients on data split by therapist subgroups defined in
each list element; see examples.}

\item{lmer}{Logical. Should the \code{\link[lme4]{lmer}} function from
packakge \code{\link[lme4]{lme4}} be used for calculating the
\code{\link[psych]{ICC}} via package \code{\link{psych}}.}

\item{verbose}{Logical, if status info should be given; defaults to TRUE.}

\item{fnOut}{Character length one. The filename (without the ending '.xlsx')
of the xlsx file where the statistics results will be saved to.}

\item{toXls}{Logical, if statistics results should be saved to xlsx.
Defaults to TRUE.}
}
\value{
A list length two, with a dataframe holding the statistics
results for the intra-class correlation coefficients in the first list
element, and the inter-class correlation coefficients in the second.
}
\description{
Takes the dataset as provided in argument 'fmddf' and calculates
intra- and interclass correlation coefficients for some pre-defined and some
customisable groupings / data-splits.
}
\details{
For calculating the resp. ICC, function \code{\link[psych]{ICC}} from
package \code{\link{psych}} is used.
}
\examples{
\dontrun{
fmdf <- importData()
res <- calcIntraInter(fmdf) # run with everything left at defaults
# additionally split one excluded muscle test, save under different name
res <- calcIntraInter(fmdf, exclOneMT=TRUE, fnOut="results_stat_MTex")
# define two therapist subgroups (in a list length two)
tsList <- list(c("Ther#1", "Ther#2", "Ther#3"), c("Ther#4", "Ther#5", "Ther#6"))
res <- calcIntraInter(fmdf, exclOneMT=FALSE, ther.sub=tsList, fnOut="results_stat_Thersub")
res
}
}
\seealso{
\code{\link{runFmdenq}}
}
