% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{importData}
\alias{importData}
\title{Import Data}
\usage{
importData(fnIn = "results_fmd", selIn = FALSE, verbose = TRUE)
}
\arguments{
\item{fnIn}{Character length one, the name of the data-input file, what is
an xlsx file residing in the folder 'results_fmd'. (The name is provided
without the '.xlsx' ending.)}

\item{selIn}{Logical, if an xlsx file should be selected interactively instead.
Defaults to FALSE.}

\item{verbose}{Logical, if the range of the values of each individual column
should be reported. Defaults to TRUE.}
}
\value{
A data frame containing FMD result data.
}
\description{
Import data from predefined xlsx sheet, check range of result
values and describe range of other values.
}
\examples{
\dontrun{
fmdf <- importData() # leave everything at the defaults
fmdf <- importData(fnIn="results_fmd_foo") # specify custom name of fmd results
# file
fmdf <- importData(sel=TRUE) # to interactively select the xlsx file to import
}
}
\seealso{
\code{\link{runFmdenq}}
}
