% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare}
\alias{prepare}
\title{Generate Folders, Copy Template}
\usage{
prepare()
}
\description{
Generate the required folders and copy a template of the xlsx
input file, all in the current working directory.
}
