% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runFmdenq}
\alias{runFmdenq}
\title{Import Data and Calculate Intra- and Interclass Correlation Coefficients}
\usage{
runFmdenq(
  fnIn = "results_fmd",
  selIn = FALSE,
  exclOneMT = FALSE,
  ther.sub = NULL,
  lmer = TRUE,
  verbose = TRUE,
  fnOut = "results_stat",
  toXls = TRUE
)
}
\arguments{
\item{fnIn}{Character length one, the name of the data-input file, what is
an xlsx file residing in the folder 'results_fmd'. (The name is provided
without the '.xlsx' ending.)}

\item{selIn}{Logical, if an xlsx file should be selected interactively instead.
Defaults to FALSE.}

\item{exclOneMT}{Logical, if additional groupings with one muscle test
excluded should be performed. Defaults to FALSE.}

\item{ther.sub}{NULL or List. If left at the default NULL, no additional
splitting with therapist subgroups is performed. Provide a list with
n therapist IDs in each list element to calculate additional inter-rater
correlation coefficients on data split by therapist subgroups defined in
each list element; see examples.}

\item{lmer}{Logical. Should the \code{\link[lme4]{lmer}} function from
packakge \code{\link[lme4]{lme4}} be used for calculating the
\code{\link[psych]{ICC}} via package \code{\link{psych}}.}

\item{verbose}{Logical, if the range of the values of each individual column
should be reported. Defaults to TRUE.}

\item{fnOut}{Character length one. The filename (without the ending '.xlsx')
of the xlsx file where the statistics results will be saved to.}

\item{toXls}{Logical, if statistics results should be saved to xlsx.
Defaults to TRUE.}
}
\description{
Imports FMD result data from a predefined xlsx file, calculates
intra- and interclass correlation coefficients on some predefined and some
optional custom data grouoings / splittings, and saves the statistics results to an
xlsx file.
}
\details{
The name of the input-excel file can be specified or it can be
interactively selected, the name of the output-excel file can be specified.
This basically is a convenient wrapper for the two functions
\code{\link{importData}} and \code{\link{calcIntraInter}}; see examples there.
}
\section{Calculating ICC}{
 For calculating the resp. ICC, function
\code{\link[psych]{ICC}} from package \code{\link{psych}} is used.
}

\examples{
\dontrun{
runFmdenq()
runFmdenq(selIn=TRUE, fnOut="results_stat_foo") # interactively select xlsx file for data import
}
}
\seealso{
\code{\link{importData}}, \code{\link{calcIntraInter}}
}
