% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{showCalcICC}
\alias{showCalcICC}
\title{Calculate ICC for Provided Data}
\usage{
showCalcICC(
  fmddf,
  expr = NULL,
  resh = NULL,
  showData = TRUE,
  showICC = TRUE,
  lmer = FALSE
)
}
\arguments{
\item{fmddf}{A reshaped data frame containing FMD data as produced by 
\code{\link{showGetData}}.}

\item{expr}{An expression defining the subset of data. If left at the default 
NULL, no subset of the data is created, but he data is merely reshaped.}

\item{resh}{Character length one. What type of reshape should be applied to 
the output. Possible values are 'inter' and 'intra'.}

\item{showData}{Logical, if the input-data should be printed as well. Defaults 
to TRUE.}

\item{showICC}{Logical, if the calculated ICC should be printed. Defaults to 
TRUE.}

\item{lmer}{Logical. Should the \code{\link[lme4]{lmer}} function from
packakge \code{\link[lme4]{lme4}} be used for calculating the
\code{\link[psych]{ICC}} via package \code{\link{psych}}.}
}
\value{
An (invisible) data frame containing the results of the ICC calculation.
}
\description{
Calculate either ICC2 or ICC3 on the provided data, possibly 
print the results and return them.
}
\section{CAVE}{
 It is the users responsibility to decide if ICC2 or ICC3 is 
appropriate for the provided data input.
}

\examples{
\dontrun{
XXX
}
}
