% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{showGetData}
\alias{showGetData}
\title{Get and Display a Subset of the Data}
\usage{
showGetData(fmddf, expr, showData = TRUE)
}
\arguments{
\item{fmddf}{A data frame containing FMD data.}

\item{expr}{An expression defining the subset of data.}

\item{showData}{Logical If the resulting data frame should be printed. Defaults 
to TRUE.}
}
\value{
A data frame with the subset of data.
}
\description{
Use an expression to provide a definition of a subset of the 
data, then possibly print these subset and return it.
}
\details{
Use 'colnames(object)' to determine the name of the columns in the 
data frame. (Of course.)
}
\examples{
\dontrun{
fmddf <- importData()
expr <- expression(Patient_ID == "Pat#2")
getShowData(fmddf, expr)
expr <- expression(Patient_ID == "Pat#2" & Therapist_ID != "Ther#4")
getShowData(fmddf, expr)
}
}
