% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ssc}
\alias{ssc}
\title{Exclude Single Class Variables}
\usage{
ssc(fmddf, cn, val, include = TRUE)
}
\arguments{
\item{fmddf}{A data frame containing FMD data.}

\item{cn}{Column name}

\item{val}{Value}

\item{include}{Logical}
}
\value{
The dataset modified
}
\description{
Define a column, define a value, and exclude or only include these 
data
}
