% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{showCalcICC}
\alias{showCalcICC}
\title{Calculate ICC for Provided Data}
\usage{
showCalcICC(
  fmddf,
  type = "ICC2",
  lmer = FALSE,
  showData = TRUE,
  showICC = TRUE
)
}
\arguments{
\item{fmddf}{A data frame containing FMD data.}

\item{type}{Character length one. Can be either 'ICC2' or 'ICC3'. Defaults to 
'ICC2'.}

\item{lmer}{Logical. Should the \code{\link[lme4]{lmer}} function from
packakge \code{\link[lme4]{lme4}} be used for calculating the
\code{\link[psych]{ICC}} via package \code{\link{psych}}.}

\item{showData}{Logical, if the input-data should be printed as well. Defaults 
to TRUE.}

\item{showICC}{Logical, if the calculated ICC should be printed. Defaults to 
TRUE.}
}
\value{
An (invisible) data frame containing the results of the ICC calculation.
}
\description{
Calculate either ICC2 or ICC3 on the provided data, possibly 
print the results and return them.
}
\section{CAVE}{
 It is the users responsibility to decide if ICC2 or ICC3 is 
appropriate for the provided data input.
}

\examples{
\dontrun{
XXX
}
}
